Uploading my 323 Assignments and Labs
I hadn't saved all of my assignments and labs so with a brief overview, there is a Term Project in here in which we created a Database for a list of airports, airlines, and airplanes to keep with various details such as flight schedules, instances, and so on and so forth. Included should also be a JDBC in which we connected Java to a Database and accessed/modified the SQL database with Java in Derby. The others are misc. files and notes and assignments done in class